Move native methods to their own object
