+++ android/src/main/java/app/edge/reactnative/core/EdgeCoreWebView.java
@@ -1,7 +1,6 @@
 package app.edge.reactnative.core;
 
 import android.graphics.Bitmap;
-import android.util.Base64;
 import android.webkit.JavascriptInterface;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
@@ -9,19 +8,12 @@ import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.uimanager.ThemedReactContext;
 import com.facebook.react.uimanager.events.RCTEventEmitter;
-import java.io.IOException;
-import java.security.SecureRandom;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
 
 class EdgeCoreWebView extends WebView {
   private static final String BASE_URL = "file:///android_asset/";
-  private final ExecutorService mPool = Executors.newCachedThreadPool();
   private final ThemedReactContext mContext;
-  private final Disklet mDisklet;
+  private final EdgeNative mNative;
 
   // react api--------------------------------------------------------------
 
@@ -47,7 +39,7 @@ class EdgeCoreWebView extends WebView {
   public EdgeCoreWebView(ThemedReactContext context) {
     super(context);
     mContext = context;
-    mDisklet = new Disklet(mContext.getFilesDir());
+    mNative = new EdgeNative(mContext.getFilesDir());
 
     getSettings().setAllowFileAccess(true);
     getSettings().setJavaScriptEnabled(true);
@@ -73,19 +65,7 @@ class EdgeCoreWebView extends WebView {
   class JsMethods {
     @JavascriptInterface
     public void call(int id, final String name, final String args) {
-      final PendingCall promise = new PendingCall(id);
-
-      mPool.execute(
-          new Runnable() {
-            @Override
-            public void run() {
-              try {
-                handleCall(name, new JSONArray(args), promise);
-              } catch (Throwable error) {
-                promise.reject(error.getMessage());
-              }
-            }
-          });
+      mNative.call(name, args, new WebViewPromise(id));
     }
 
     @JavascriptInterface
@@ -107,56 +87,6 @@ class EdgeCoreWebView extends WebView {
 
   // utilities -------------------------------------------------------------
 
-  private void handleCall(String name, JSONArray args, PendingCall promise)
-      throws IOException, JSONException {
-    switch (name) {
-      case "diskletDelete":
-        mDisklet.delete(args.getString(0));
-        promise.resolve(null);
-        break;
-      case "diskletGetData":
-        promise.resolve(Base64.encodeToString(mDisklet.getData(args.getString(0)), Base64.NO_WRAP));
-        break;
-      case "diskletGetText":
-        promise.resolve(mDisklet.getText(args.getString(0)));
-        break;
-      case "diskletList":
-        promise.resolve(new JSONObject(mDisklet.list(args.getString(0))));
-        break;
-      case "diskletSetData":
-        mDisklet.setData(args.getString(0), Base64.decode(args.getString(1), Base64.DEFAULT));
-        promise.resolve(null);
-        break;
-      case "diskletSetText":
-        mDisklet.setText(args.getString(0), args.getString(1));
-        promise.resolve(null);
-        break;
-      case "randomBytes":
-        {
-          SecureRandom sr = new SecureRandom();
-          byte[] entropy = new byte[args.getInt(0)];
-          sr.nextBytes(entropy);
-          promise.resolve(Base64.encodeToString(entropy, Base64.NO_WRAP));
-        }
-        break;
-      case "scrypt":
-        {
-          byte[] data = Base64.decode(args.getString(0), Base64.DEFAULT);
-          byte[] salt = Base64.decode(args.getString(1), Base64.DEFAULT);
-          int n = args.getInt(2);
-          int r = args.getInt(3);
-          int p = args.getInt(4);
-          int dklen = args.getInt(5);
-          byte[] out = scrypt(data, salt, n, r, p, dklen);
-          if (out == null) promise.reject("Failed scrypt");
-          else promise.resolve(Base64.encodeToString(out, Base64.NO_WRAP));
-        }
-        break;
-      default:
-        promise.reject("No method " + name);
-    }
-  }
-
   private void visitPage() {
     String source = mSource == null ? BASE_URL + "edge-core-js/edge-core.js" : mSource;
     String html =
@@ -172,17 +102,19 @@ class EdgeCoreWebView extends WebView {
     loadDataWithBaseURL(BASE_URL, html, "text/html", null, null);
   }
 
-  private class PendingCall {
+  private class WebViewPromise implements PendingCall {
     private final int mId;
 
-    PendingCall(int id) {
+    WebViewPromise(int id) {
       mId = id;
     }
 
+    @Override
     public void resolve(Object value) {
       runJs("window.nativeBridge.resolve(" + mId + "," + stringify(value) + ")");
     }
 
+    @Override
     public void reject(String message) {
       runJs("window.nativeBridge.reject(" + mId + "," + stringify(message) + ")");
     }
@@ -195,11 +127,5 @@ class EdgeCoreWebView extends WebView {
           .replace("\u2028", "\\u2028")
           .replace("\u2029", "\\u2029");
     }
-  }
-
-  private native byte[] scrypt(byte[] data, byte[] salt, int n, int r, int p, int dklen);
-
-  static {
-    System.loadLibrary("edge-core-jni");
   }
 }
