+++ android/src/main/java/app/edge/reactnative/core/PendingCall.java
@@ -1,0 +1,7 @@
+package app.edge.reactnative.core;
+
+public interface PendingCall {
+  public void resolve(Object value);
+
+  public void reject(String message);
+}
