+++ ios/PendingCall.swift
@@ -1,0 +1,4 @@
+struct PendingCall {
+  var resolve: (_ value: Any?) -> Void
+  var reject: (_ message: String) -> Void
+}
